Abstract Class
The Same Method that Perform different tasks according to the object calling it. 
A Method With no body should be implemented in it's Sub classes.
It's Used when there are common Features shared by all the Objects.
A class with Abstract Method is Called an Abstract Class.
An abstract class can Have instance variables and concrete Methods.