Inner Classes.
Classes that are nested together are known as Inner Classes.
Purpose of inner classes is to group together classes that are related. 
Inner classes have full access to the members of the outer class even if they're private.

To access the members of an inner class, you first create an object to its outer class. Then use that reference to create an object of the inner class.
Example:
        OuterClass outerC = new OuterClass;
        OuterClass.InnerClass inner = outerC.new OuterClass.InnerClass;

If you don't Want the outer classes to access your inner class then you can make it Private/Protected.

Local Inner class:
-Declared inside the Main.
-Useful when dealing with a class with many objects.

Anonymous inner Class:
-Doesn't have a name.
-Needs to be declared and instantiated at the same time.
-Useful when dealing with an object instance that you'd use only once.